requirements.txt
-i https://pypi.org/simple
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
iniconfig==2.1.0; python_version >= '3.8'
packaging==25.0; python_version >= '3.8'
pluggy==1.6.0; python_version >= '3.9'
pygments==2.19.2; python_version >= '3.8'
pytest==8.4.2; python_version >= '3.9'
blinker==1.9.0; python_version >= '3.9'
click==8.3.0; python_version >= '3.10'
flask==3.1.2; python_version >= '3.9'
flask-login==0.6.3; python_version >= '3.7'
flask-wtf==1.2.2; python_version >= '3.9'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
markupsafe==3.0.2; python_version >= '3.9'
psycopg2-binary==2.9.10; python_version >= '3.8'
python-dotenv==1.1.1; python_version >= '3.9'
werkzeug==3.1.3; python_version >= '3.9'
wtforms==3.2.1; python_version >= '3.9'
